voice arduino bridge node first draft
